rl_training/requirements.txt
torch>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
setuptools<81
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0
pandas>=2.0.0
streamlit>=1.32.0
openai>=1.0.0
anthropic>=0.18.0
